Mohsen Ahmadi - Amirhossein Mazaheri/Q4-5/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
dj-database-url
celery
redis
httpx
pytest
pytest-asyncio

djangorestframework
djangorestframework-simplejwt
django-environ
dj-database-url
docker